+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# cli development version
+
 # cli 2.5.0
 
 * New `style_no_*()` functions to locally undo styling.