+++ NEWS.md
@@ -16,6 +16,9 @@
 * `cli_memo()` was renamed to `cli_bullets()`, as it is by default
   formatted as a bullet list (#250).
 
+* New `cli_abort()`, `cli_warn()` and `cli_inform()` functions, to throw
+  errors with cli pluralization and styling.
+
 * New `ansi_toupper()`, `ansi_tolower` and `ansi_chartr()` functions,
   the ANSI styling aware variants of `toupper()`, `tolower()` and
   `chartr()` (#248).